loto-predictor/requirements.txt
tensorflow==2.10.0
geneticalgorithm==1.0.2
numpy==1.22.4
pandas==1.5.3
scikit-learn==1.2.2
matplotlib==3.7.1
pyyaml==6.0
tqdm==4.65.0